PIC now goes to sleep when no buttons are pushed.
Now when no buttons are pushed, the MCU will go to sleep. I can confirm
that this has decreased power consumption substantially. Without the
added call to Sleep(), current draw was approx 5mA. Now it is at 1mA. I
tried a call to Idle() as well, but that only managed to decrease
current to around 2.5mA.