New Branch with Nestor Shenanigan
Trying to refactor our logic to be outside of main.c

Added flash_led.c and uart_com.c

As of right now this new version of the code works in almost all cases.
However, if one is "trying" to break it (pressing or unpressing 2
buttons in unison repeatedly) it can get stuck in a state. Takes about
10 tries. Some more debouncing is needed.